@@ -1,4 +1,4 @@
-class User < ActiveRecord::Base
+class User < ApplicationRecord
   scope :active, lambda { where(active: true) }
   scope :inactive, lambda { where(active: false) }
 end
