@@ -1,88 +1,98 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe MailyHerald::Utils do
+
   describe MailyHerald::Utils::MarkupEvaluator do
-    before(:each) do
-      @mailing = MailyHerald.one_time_mailing :test_mailing
-      @user = FactoryGirl.create :user
-      @list = @mailing.list
-      @list.subscribe!(@user)
-      @subscription = @mailing.subscription_for(@user)
-      @evaluator = MailyHerald::Utils::MarkupEvaluator.new(@list.context.drop_for(@user, @subscription))
-    end
 
-    describe "conditions" do
-      it "should validate syntax" do
-        expect(@mailing).to be_valid
-        @mailing.conditions = "foo bar"
-        #expect {@evaluator.evaluate_conditions(@mailing.conditions)}.to raise_error(Liquid::Error)
-        expect(MailyHerald::Utils::MarkupEvaluator.test_conditions(@mailing.conditions)).to be_falsy
-        expect(@mailing).not_to be_valid
-        expect(@mailing.errors).to include(:conditions)
+    let!(:entity) { create :user }
+    let!(:mailing) { create :generic_one_time_mailing }
+    let(:list) { mailing.list }
+    let(:subscription) { mailing.subscription_for entity }
+    let(:evaluator) { described_class.new(list.context.drop_for(entity, subscription)) }
+
+    before { list.subscribe! entity }
+
+    context "conditions" do
+      it { expect(mailing).to be_valid }
+
+      context "syntax" do
+        before { mailing.conditions = "foo bar" }
+
+        it { expect(described_class.test_conditions(mailing.conditions)).to be_falsy }
+        it { expect(mailing).not_to be_valid }
+        it { mailing.valid?; expect(mailing.errors).to include(:conditions) }
       end
 
-      it "should validate numerical conditions" do
-        @mailing.conditions = "2"
-        expect {@mailing.conditions_met?(@user)}.to raise_error
+      context "numerical" do
+        context "integer" do
+          before { mailing.conditions = "2" }
+          it { expect {mailing.conditions_met?(entity)}.to raise_error(ArgumentError) }
+        end
 
-        @mailing.conditions = "5 == 5"
-        @evaluator.evaluate_conditions(@mailing.conditions).should be_truthy
-        expect {@mailing.conditions_met?(@user)}.not_to raise_error
+        context "true condition" do
+          before { mailing.conditions = "5 == 5" }
+          it { expect(evaluator.evaluate_conditions(mailing.conditions)).to be_truthy }
+          it { expect {mailing.conditions_met?(entity)}.not_to raise_error }
+        end
 
-        @mailing.conditions = "2 == 3"
-        @evaluator.evaluate_conditions(@mailing.conditions).should be_falsy
-        expect {@mailing.conditions_met?(@user)}.not_to raise_error
+        context "false condition" do
+          before { mailing.conditions = "2 == 3" }
+          it { expect(evaluator.evaluate_conditions(mailing.conditions)).to be_falsy }
+          it { expect {mailing.conditions_met?(entity)}.not_to raise_error }
+        end
       end
 
-      it "should evaluate more complex conditions" do
-        expect(@user.weekly_notifications).to be_truthy
+      context "more complex" do
+        it { expect(entity.weekly_notifications).to be_truthy }
+      
+        context "check if true" do
+          before { mailing.conditions = "user.weekly_notifications == true" }
+          it { expect(evaluator.evaluate_conditions(mailing.conditions)).to be_truthy }
+          it { expect {mailing.conditions_met?(entity)}.not_to raise_error }
+        end
+      
+        context "check if false" do
+          before { mailing.conditions = "user.weekly_notifications == false" }
+          it { expect(evaluator.evaluate_conditions(mailing.conditions)).to be_falsy }
+          it { expect {mailing.conditions_met?(entity)}.not_to raise_error }
+        end
 
-        @mailing.conditions = "user.weekly_notifications == true"
-        @evaluator.evaluate_conditions(@mailing.conditions).should be_truthy
-        expect {@mailing.conditions_met?(@user)}.not_to raise_error
+        context "update to false and check if true" do
+          before do
+            entity.update_attributes! weekly_notifications: false
+            expect(entity.weekly_notifications).to be_falsy
+            mailing.conditions = "user.weekly_notifications == true"
+          end
 
-        @mailing.conditions = "user.weekly_notifications == false"
-        @evaluator.evaluate_conditions(@mailing.conditions).should be_falsy
-        expect {@mailing.conditions_met?(@user)}.not_to raise_error
-
-        @user.weekly_notifications = false
-        expect(@user.weekly_notifications).to be_falsy
-
-        @mailing.conditions = "user.weekly_notifications == true"
-        @evaluator.evaluate_conditions(@mailing.conditions).should be_falsy
-        expect {@mailing.conditions_met?(@user)}.not_to raise_error
-      end
-
-      it "should provide model conditions syntax validation" do
-        @mailing.should be_valid
-        @mailing.conditions = "foo bar"
-        @mailing.should_not be_valid
-      end
-    end
-
-    describe "start at" do
-      it "should validate syntax" do
-        expect(@mailing).to be_valid
-        @mailing.start_at = "- foo bar | ddd"
-        expect(MailyHerald::Utils::MarkupEvaluator.test_start_at(@mailing.start_at)).to be_falsy
-        expect(@mailing).not_to be_valid
-        expect(@mailing.errors).to include(:start_at)
-      end
-
-      it "should evaluate attributes" do
-        expect(@evaluator.evaluate_start_at("user.created_at")).to eq(@user.created_at)
-      end
-
-      it "should evaluate attributes with filters" do
-        expect(@evaluator.evaluate_start_at("user.created_at | minus: 1, 'day'")).to eq(@user.created_at - 1.day)
-
-        expect(@evaluator.evaluate_start_at("user.created_at | plus: 2, 'minutes'")).to eq(@user.created_at + 2.minutes)
-      end
-
-      it "should evaluate attributes without subscription" do
-        @evaluator = MailyHerald::Utils::MarkupEvaluator.new(@list.context.drop_for(@user, nil))
-        expect(@evaluator.evaluate_start_at("user.created_at")).to eq(@user.created_at)
+          it { expect(evaluator.evaluate_conditions(mailing.conditions)).to be_falsy }
+          it { expect {mailing.conditions_met?(entity)}.not_to raise_error }
+        end
       end
     end
+
+    context "start_at" do
+      context "validate syntax" do
+        before { mailing.start_at = "- foo bar | ddd" }
+          it { expect(described_class.test_start_at(mailing.start_at)).to be_falsy }
+          it { expect(mailing).not_to be_valid }
+          it { mailing.valid?; expect(mailing.errors).to include(:start_at) }
+      end
+
+      context "evaluate without filters" do
+        it { expect(evaluator.evaluate_start_at("user.created_at")).to eq(entity.created_at) }
+      end
+
+      context "evaluate with filters" do
+        it { expect(evaluator.evaluate_start_at("user.created_at | minus: 1, 'day'")).to eq(entity.created_at - 1.day) }
+        it { expect(evaluator.evaluate_start_at("user.created_at | plus: 2, 'minutes'")).to eq(entity.created_at + 2.minutes) }
+      end
+
+      context "evaluate without subscription" do
+        let!(:evaluator) { described_class.new(list.context.drop_for(entity, nil)) }
+        it { expect(evaluator.evaluate_start_at("user.created_at")).to eq(entity.created_at) }
+      end
+    end
+
   end
+
 end
