@@ -4,7 +4,7 @@ module MailyHerald
     validates   :start_at,      presence: true
     validate    :validate_start_at
 
-    after_save :update_schedules_callback, if: Proc.new{|m| m.state_changed? || m.start_at_changed? || m.override_subscription?}
+    after_save :update_schedules_callback, if: Proc.new{|m| m.saved_change_to_attribute?(:state) || m.saved_change_to_attribute?(:start_at) || m.override_subscription?}
 
     # Sends mailing to all subscribed entities.
     #
