@@ -3,8 +3,8 @@ module MailyHerald
     isolate_namespace MailyHerald
 
     config.generators do |g|
-      g.test_framework      :rspec,         fixture: false
-      g.fixture_replacement :factory_girl,  dir: 'spec/factories'
+      g.test_framework :rspec
+      g.fixture_replacement :factory_girl, :dir => 'spec/support/factories'
     end
 
     config.to_prepare do
