@@ -1,3 +1,3 @@
 module MailyHerald
-  VERSION = "0.9.4"
+  VERSION = '1.0.0'
 end
