@@ -1,81 +1,33 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe MailyHerald::TokensController do
-  before(:each) do
-    @user = FactoryGirl.create :user
-    @mailing = MailyHerald.one_time_mailing :test_mailing
-    @subscription = @mailing.subscription_for @user
-  end
 
-  describe "Unsubscribe action" do
-    before(:each) do
-      @mailing.token_action = :unsubscribe
-      @mailing.save
+  routes { MailyHerald::Engine.routes }
+
+  let!(:entity) { create :user }
+  let!(:mailing) { create :generic_one_time_mailing }
+  let(:subscription) { mailing.subscription_for entity }
+
+  before { mailing.list.subscribe! entity }
+
+  it { expect(subscription.active?).to be_truthy }
+
+  describe "GET #get" do
+    context "with valid token" do
+      before { get :get, params: {token: subscription.token} }
+
+      it { subscription.reload; expect(subscription.active?).to be_falsy }
+      it { expect(response).to redirect_to('/') }
+      it { expect(flash[:notice]).to eq(I18n.t('maily_herald.subscription.deactivated')) }
     end
 
-    describe "when regular subscription" do
-      pending "should deactivate only one subscription" do
-        get :get, token: @subscription.token, use_route: :maily_herald
-        expect(response).to redirect_to("/")
-        @subscription.reload
+    context "with invalid token" do
+      before { get :get, params: {token: "invalid_token"} }
 
-        expect(@subscription.active?).not_to be_true
-
-        @user.maily_herald_subscriptions.each do |s|
-          next unless s.target.subscription_group == @subscription.target.subscription_group
-          next if s == @subscription
-
-          expect(s.active?).to be_true
-        end
-      end
-    end
-
-    describe "when aggregated subscription" do
-      before(:each) do
-        @mailing.subscription_group = :account
-        @mailing.save!
-      end
-
-      after(:each) do
-        @mailing.subscription_group = nil
-        @mailing.save!
-      end
-
-      pending "should deactivate subscription group" do
-        get :get, token: @subscription.token, use_route: :maily_herald
-        expect(response).to redirect_to("/")
-        @subscription.reload
-
-        expect(@subscription.active?).not_to be_true
-        expect(@subscription.aggregate).not_to be_nil
-        expect(@subscription.aggregate.active?).not_to be_true
-
-        @user.maily_herald_subscriptions.each do |s|
-          next unless s.target.subscription_group == @subscription.target.subscription_group
-
-          expect(s.active?).to be_false
-        end
-      end
+      it { subscription.reload; expect(subscription.active?).to be_truthy }
+      it { expect(response).to redirect_to('/') }
+      it { expect(flash[:notice]).to eq(I18n.t('maily_herald.subscription.undefined_token')) }
     end
   end
 
-  pending "Custom action" do
-    before(:each) do
-      @mailing.token_action = :custom
-      expect(@mailing).to be_valid
-      @mailing.save
-      expect(@mailing.token_custom_action).not_to be_nil
-    end
-
-    pending "should perform custom action" do
-      @subscription.reload
-      expect(@subscription.target.token_action).to eq(:custom)
-      get :get, token: @subscription.token, use_route: :maily_herald
-      expect(response).to redirect_to("/custom")
-      @subscription.reload
-      @user.reload
-
-      expect(@user.name).to eq("changed")
-    end
-  end
 end
