@@ -1,6 +1,4 @@
-class AdHocMailer < MailyHerald::Mailer
-  default from: "no-reply@mailyherald.org"
-
+class AdHocMailer < ApplicationMailer
   def ad_hoc_mail user
     mail to: user.email, subject: "Test"
   end
