@@ -1,41 +1,26 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe MailyHerald::Context do
-  describe "setup" do
-    before(:each) do
-      @user = FactoryGirl.create :user
-      @mailing = MailyHerald.one_time_mailing :test_mailing
-      @list = @mailing.list
-      @context = @list.context
-    end
 
-    describe "with subscription" do
-      before(:each) do
-        @list.subscribe! @user
-        @subscription = @mailing.subscription_for @user
-        @drop = @context.drop_for @user, @subscription
-      end
+  let!(:entity) { create :user }
+  let!(:mailing) { create :generic_one_time_mailing }
+  let!(:list) { mailing.list }
+  let!(:context) { list.context }
 
-      it "should get valid context" do
-        expect(@context).to be_kind_of(MailyHerald::Context)
-      end
+  it { expect(context.destination_for(entity)).to eq(entity.email) }
+  it { expect(context.destination_attribute).to be_nil }
 
-      it "should resolve attributes properly" do
-        expect(@drop["user"]).to be_kind_of(MailyHerald::Context::Drop)
-        expect(@drop["user"]["name"]).to eq(@user.name)
-        expect(@drop["user"]["properties"]["prop1"]).to eq(@user.name[0])
-      end
+  context "with subscription" do
+    let(:subscription) { mailing.subscription_for entity }
+    let(:drop) { context.drop_for entity, subscription }
 
-      it "should resolve subscription attributes properly" do
-        expect(@drop["subscription"]).to be_kind_of(MailyHerald::Subscription)
-      end
-    end
+    before { list.subscribe! entity }
+
+    it { expect(context).to be_kind_of(MailyHerald::Context) }
+    it { expect(drop["user"]).to be_kind_of(MailyHerald::Context::Drop) }
+    it { expect(drop["user"]["name"]).to eq(entity.name) }
+    it { expect(drop["user"]["properties"]["prop1"]).to eq(entity.name[0]) }
+    it { expect(drop["subscription"]).to be_kind_of(MailyHerald::Subscription) }
   end
 
-  it "should handle both destination procs and strings" do
-    @user = FactoryGirl.create :user
-    context = MailyHerald.context :all_users
-    expect(context.destination_for(@user)).to eq(@user.email)
-    expect(context.destination_attribute).to be_nil
-  end
 end
