@@ -1,285 +1,247 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe MailyHerald::OneTimeMailing do
-  before(:each) do
-    @entity = FactoryGirl.create :user
 
-    @list = MailyHerald.list(:generic_list)
-    expect(@list.context).to be_a(MailyHerald::Context)
-  end
+  let!(:entity) { create :user }
+  let!(:list) { MailyHerald.list :generic_list }
 
-  after(:all) do
-    Timecop.return
-  end
+  after(:all) { Timecop.return }
 
-  describe "with subscription" do
-    before(:each) do
-      @list.subscribe!(@entity)
+  it { expect(list.context).to be_a(MailyHerald::Context) }
+
+  context "with subscription" do
+    before { list.subscribe!(entity) }
+
+    context "run all delivery" do
+      let!(:mailing) { create :generic_one_time_mailing }
+
+      it { expect(mailing).to be_kind_of(MailyHerald::OneTimeMailing) }
+      it { expect(mailing).not_to be_a_new_record }
+      it { expect(mailing).to be_valid }
+
+      context "delivery" do
+        let!(:subscription) { mailing.subscription_for(entity) }
+
+        it { expect(MailyHerald::Subscription.count).to eq(1) }
+        it { expect(MailyHerald::Log.delivered.count).to eq(0) }
+
+        it { expect(subscription).to be_kind_of(MailyHerald::Subscription) }
+
+        it { expect(mailing.logs.scheduled.count).to eq(1) }
+        it { expect(mailing.conditions_met?(entity)).to be_truthy }
+        it { expect(mailing.processable?(entity)).to be_truthy }
+        it { expect(mailing.mailer_name).to eq(:generic) }
+        it { expect(mailing.schedules.for_entity(entity).count).to eq(1) }
+
+        context "run" do
+          let(:ret) { mailing.run }
+
+          it { expect(ret).to be_kind_of(Array) }
+          it { expect(ret.first).to be_kind_of(MailyHerald::Log) }
+          it { expect(ret.first).to be_delivered }
+          it { expect(ret.first.mail).to be_kind_of(Mail::Message) }
+
+          context "after run" do
+            before { ret; mailing.set_schedules }
+
+            it { expect(MailyHerald::Subscription.count).to eq(1) }
+            it { expect(MailyHerald::Log.delivered.count).to eq(1) }
+
+            it { expect(mailing.logs.processed.for_entity(entity).count).to eq(1) }
+            it { expect(mailing.schedules.for_entity(entity).count).to eq(0) }
+
+            it "log should have proper values" do
+              log = MailyHerald::Log.delivered.first
+              expect(log.entity).to eq(entity)
+              expect(log.mailing).to eq(mailing)
+              expect(log.entity_email).to eq(entity.email)
+            end
+
+            context "runnning again" do
+              before { ret }
+
+              it { expect(mailing.logs.processed.for_entity(entity).count).to eq(1) }
+              it { expect(mailing.schedules.for_entity(entity).count).to eq(0) }
+            end
+          end
+        end
+      end
+
+      context "template errors", raise_delivery_errors: false do
+        let!(:mailing) { create :mail_with_error }
+        let!(:schedule) { mailing.schedule_for(entity) }
+
+        before do
+          expect(MailyHerald::Log.delivered.count).to eq(0)
+          expect(schedule).to be_a(MailyHerald::Log)
+          expect(schedule.processing_at).to be <= Time.now
+          mailing.run
+          schedule.reload
+        end
+
+        it { expect(schedule).to be_error }
+      end
     end
 
-    describe "run all delivery" do
-      before(:each) do
-        @mailing = MailyHerald.one_time_mailing(:test_mailing)
-        expect(@mailing).to be_kind_of(MailyHerald::OneTimeMailing)
-        expect(@mailing).not_to be_a_new_record
-        expect(@mailing).to be_valid
+    context "single entity delivery" do
+      let!(:mailing)  { create :custom_one_time_mailing }
+      let!(:schedule) { mailing.schedule_for(entity) }
+      let(:mailer)    { mailing.mailer }
+
+      before { schedule.update_attribute(:processing_at, Time.now + 1.day) }
+
+      it { expect(mailing.logs.delivered.count).to eq(0) }
+      it { expect{ mailer.one_time_mail(entity).deliver }.not_to change{ActionMailer::Base.deliveries.count} }
+    end
+
+    context "with entity outside the scope" do
+      let!(:mailing) { create :generic_one_time_mailing }
+
+      context "check setup - active" do
+        it { expect(list.context.scope).to include(entity) }
+        it { expect(mailing).to be_processable(entity) }
+        it { expect(mailing).to be_enabled }
       end
 
-      it "should be delivered only once per user" do
-        subscription = @mailing.subscription_for(@entity)
+      context "after setup - non active" do
+        before { entity.update_attributes!(active: false) }
 
-        expect(@mailing.logs.scheduled.count).to eq(1)
-        expect(@mailing.schedules.for_entity(@entity).count).to eq(1)
+        it { expect(list.context.scope).not_to include(entity) }
+        it { expect(list).to be_subscribed(entity) }
+        it { expect(mailing).not_to be_processable(entity) }
 
-        ret = @mailing.run
+        it "should not process mailings, postpone them and finally skip them" do
+          schedule = mailing.schedule_for(entity)
+          processing_at = schedule.processing_at
+          expect(schedule).not_to be_nil
+          expect(schedule.processing_at).to be <= Time.now
+          
+          mailing.run
 
-        expect(@mailing.logs.processed.for_entity(@entity).count).to eq(1)
-        expect(@mailing.schedules.for_entity(@entity).count).to eq(0)
+          schedule.reload
+          expect(schedule).to be_scheduled
+          expect(schedule.processing_at.to_i).to eq((Time.now + 1.day).to_i)
+          expect(schedule.data[:original_processing_at]).to eq(processing_at)
+          expect(schedule.data[:delivery_attempts].length).to eq(1)
 
-        @mailing.set_schedules
+          Timecop.freeze schedule.processing_at + 1
 
-        expect(@mailing.schedules.for_entity(@entity).count).to eq(0)
+          mailing.run
 
-        ret = @mailing.run
+          schedule.reload
+          expect(schedule).to be_scheduled
+          expect(schedule.data[:delivery_attempts].length).to eq(2)
 
-        expect(@mailing.logs.processed.for_entity(@entity).count).to eq(1)
-        expect(@mailing.schedules.for_entity(@entity).count).to eq(0)
+          Timecop.freeze schedule.processing_at + 1
+
+          mailing.run
+
+          schedule.reload
+          expect(schedule).to be_scheduled
+          expect(schedule.data[:delivery_attempts].length).to eq(3)
+
+          Timecop.freeze schedule.processing_at + 1
+
+          mailing.run
+
+          schedule.reload
+          expect(schedule).to be_skipped
+          expect(schedule.data[:delivery_attempts].length).to eq(3)
+          expect(schedule.data[:skip_reason]).to eq(:not_in_scope)
+        end
       end
+    end
+  end
+
+  context "with subscription override" do
+    let!(:mailing)  { create :custom_one_time_mailing }
+
+    before { mailing.update_attributes!(override_subscription: true) }
+    after  { mailing.update_attributes!(override_subscription: false) }
+
+    it { expect(MailyHerald::Log.delivered.count).to eq(0) }
+    it { expect(mailing.processable?(entity)).to be_truthy }
+    it { expect(mailing.override_subscription?).to be_truthy }
+    it { expect(mailing.enabled?).to be_truthy }
+    it { mailing.run; expect(MailyHerald::Log.delivered.count).to eq(1) }
+  end
+
+  context "with block start_at" do
+    # FIXME: Set the id manually so it doesn't interfere with other mailings that may have the same id during tests but no 'start_at' proc.
+    let!(:mailing) { create :custom_one_time_mailing, id: 99, start_at: Proc.new{|user| user.created_at + 1.hour} }
+
+    it { expect(mailing.has_start_at_proc?).to be_truthy }
+    it { expect(mailing.processed_logs(entity).count).to eq(0) }
+    it { expect(mailing.schedules.for_entity(entity).count).to eq(0) }
+
+    context "after subscribing" do
+      before { list.subscribe!(entity) }
+
+      it { expect(list.subscription_for(entity)).to be_a(MailyHerald::Subscription) }
+      it { expect(list.subscription_for(entity)).to be_active }
+      it { expect(list.subscribed?(entity)).to be_truthy }
+
+      context "automatic schedule updater should be triggered" do
+        it { expect(mailing.schedules.for_entity(entity).count).to eq(1) }
+        it { expect(mailing.schedules.for_entity(entity).last.processing_at.to_i).to eq((entity.created_at + 1.hour).to_i) }
+      end
+
+      context "manually setting schedules should not change anything now" do
+        before { mailing.set_schedules }
+
+        it { expect(mailing.schedules.for_entity(entity).count).to eq(1) }
+        it { expect(mailing.schedules.for_entity(entity).last.processing_at.to_i).to eq((entity.created_at + 1.hour).to_i) }
+      end
+    end
+  end
+
+  context "with block conditions" do
+    let!(:mailing) { create :custom_one_time_mailing, conditions: Proc.new {|user| user.weekly_notifications} }
+
+    it { expect(mailing.has_conditions_proc?).to be_truthy }
+
+    context "when positive" do
+      let!(:entity) { create :user, weekly_notifications: true }
+      let(:schedule) { mailing.schedules.for_entity(entity).last }
+
+      before do
+        list.subscribe! entity
+        mailing.set_schedule_for entity
+      end
+
+      it { expect(mailing.schedules.for_entity(entity).count).to eq(1) }
+      it { expect(schedule.processing_at.to_i).to eq(entity.created_at.to_i) }
+      it { expect(entity.weekly_notifications).to be_truthy }
+      it { expect(mailing.conditions_met?(entity)).to be_truthy }
 
       it "should be delivered" do
-        subscription = @mailing.subscription_for(@entity)
-
-        expect(MailyHerald::Subscription.count).to eq(1)
-        expect(MailyHerald::Log.delivered.count).to eq(0)
-        expect(@mailing.logs.scheduled.count).to eq(1)
-
-        expect(subscription).to be_kind_of(MailyHerald::Subscription)
-
-        expect(@mailing.conditions_met?(@entity)).to be_truthy
-        expect(@mailing.processable?(@entity)).to be_truthy
-        expect(@mailing.mailer_name).to eq(:generic)
-
-        ret = @mailing.run
-        expect(ret).to be_kind_of(Array)
-        expect(ret.first).to be_kind_of(MailyHerald::Log)
-        expect(ret.first).to be_delivered
-        expect(ret.first.mail).to be_kind_of(Mail::Message)
-
-        expect(MailyHerald::Subscription.count).to eq(1)
-        expect(MailyHerald::Log.delivered.count).to eq(1)
-
-        log = MailyHerald::Log.delivered.first
-        expect(log.entity).to eq(@entity)
-        expect(log.mailing).to eq(@mailing)
-        expect(log.entity_email).to eq(@entity.email)
-      end
-
-      it "should handle template errors" do
-        MailyHerald::Log.delivered.count.should eq(0)
-
-        @mailing = MailyHerald.dispatch(:mail_with_error)
-        schedule = @mailing.schedule_for(@entity)
-        expect(schedule).to be_a(MailyHerald::Log)
-        expect(schedule.processing_at).to be <= Time.now
-
-        @mailing.run
-
+        schedule = mailing.schedules.for_entity(entity).last
+        mailing.run
         schedule.reload
-        expect(schedule).to be_error
+        expect(schedule.status).to eq(:delivered)
       end
     end
 
-    describe "single entity delivery" do
-      it "should not be possible via Mailer" do
-        expect(MailyHerald::Log.delivered.count).to eq(0)
+    context "when negative" do
+      let!(:entity) { create :user, weekly_notifications: false }
 
-        schedule = MailyHerald.dispatch(:one_time_mail).schedule_for(@entity)
-        schedule.update_attribute(:processing_at, Time.now + 1.day)
-
-        expect{ CustomOneTimeMailer.one_time_mail(@entity).deliver }.not_to change{ActionMailer::Base.deliveries.count}
-
-        expect(MailyHerald::Log.delivered.count).to eq(0)
-      end
-    end
-
-    describe "with entity outside the scope" do
-      before(:each) do
-        @mailing = MailyHerald.one_time_mailing(:test_mailing)
+      before do
+        list.subscribe! entity
+        mailing.set_schedules
       end
 
-      it "should not process mailings, postpone them and finally skip them" do
-        expect(@list.context.scope).to include(@entity)
-        expect(@mailing).to be_processable(@entity)
-        expect(@mailing).to be_enabled
+      it { expect(mailing.schedules.for_entity(entity).count).to eq(1) }
+      it { expect(mailing.schedules.for_entity(entity).last.processing_at.to_i).to eq(entity.created_at.to_i) }
+      it { expect(entity.weekly_notifications).to be_falsey }
+      it { expect(mailing.conditions_met?(entity)).to be_falsey }
 
-        @entity.update_attribute(:active, false)
-
-        expect(@list.context.scope).not_to include(@entity)
-        expect(@list).to be_subscribed(@entity)
-
-        expect(@mailing).not_to be_processable(@entity)
-
-        schedule = @mailing.schedule_for(@entity)
-        processing_at = schedule.processing_at
-        expect(schedule).not_to be_nil
-        expect(schedule.processing_at).to be <= Time.now
-        
-        @mailing.run
-
+      it "should be skipped" do
+        schedule = mailing.schedules.for_entity(entity).last
+        mailing.run
         schedule.reload
-        expect(schedule).to be_scheduled
-        expect(schedule.processing_at.to_i).to eq((Time.now + 1.day).to_i)
-        expect(schedule.data[:original_processing_at]).to eq(processing_at)
-        expect(schedule.data[:delivery_attempts].length).to eq(1)
-
-        Timecop.freeze schedule.processing_at + 1
-
-        @mailing.run
-
-        schedule.reload
-        expect(schedule).to be_scheduled
-        expect(schedule.data[:delivery_attempts].length).to eq(2)
-
-        Timecop.freeze schedule.processing_at + 1
-
-        @mailing.run
-
-        schedule.reload
-        expect(schedule).to be_scheduled
-        expect(schedule.data[:delivery_attempts].length).to eq(3)
-
-        Timecop.freeze schedule.processing_at + 1
-
-        @mailing.run
-
-        schedule.reload
-        expect(schedule).to be_skipped
-        expect(schedule.data[:delivery_attempts].length).to eq(3)
-        expect(schedule.data[:skip_reason]).to eq(:not_in_scope)
+        expect(schedule.status).to eq(:skipped)
       end
     end
   end
 
-  describe "with subscription override" do
-    before(:each) do
-      @mailing = MailyHerald.one_time_mailing(:one_time_mail)
-      @mailing.update_attribute(:override_subscription, true)
-    end
-
-    after(:each) do
-      @mailing.update_attribute(:override_subscription, false)
-    end
-
-    it "should deliver single mail" do
-      expect(MailyHerald::Log.delivered.count).to eq(0)
-      expect(@mailing.processable?(@entity)).to be_truthy
-      expect(@mailing.override_subscription?).to be_truthy
-      expect(@mailing.enabled?).to be_truthy
-      @mailing.run
-      expect(MailyHerald::Log.delivered.count).to eq(1)
-    end
-  end
-
-  describe "with block start_at" do
-    before(:each) do
-      @mailing = MailyHerald::OneTimeMailing.new
-      @mailing.title = "Foobar"
-      @mailing.subject = "Foo"
-      @mailing.template = "Sample template"
-      @mailing.list = @list
-      @mailing.start_at = Proc.new{|user| user.created_at + 1.hour}
-      @mailing.enable
-      @mailing.save!
-    end
-
-    after(:each) do
-      @mailing.destroy
-    end
-
-    it "should be delivered" do
-      expect(@mailing.has_start_at_proc?).to be_truthy
-
-      expect(@mailing.processed_logs(@entity).count).to eq(0)
-      expect(@mailing.schedules.for_entity(@entity).count).to eq(0)
-
-      @list.subscribe!(@entity)
-
-      expect(@list.subscription_for(@entity)).to be_a(MailyHerald::Subscription)
-      expect(@list.subscription_for(@entity)).to be_active
-      expect(@list.subscribed?(@entity)).to be_truthy
-
-      # automatic schedule updater should be triggered
-      expect(@mailing.schedules.for_entity(@entity).count).to eq(1)
-      expect(@mailing.schedules.for_entity(@entity).last.processing_at.to_i).to eq((@entity.created_at + 1.hour).to_i)
-
-      # manually setting schedules should not change anything now
-      @mailing.set_schedules
-
-      expect(@mailing.schedules.for_entity(@entity).count).to eq(1)
-      expect(@mailing.schedules.for_entity(@entity).last.processing_at.to_i).to eq((@entity.created_at + 1.hour).to_i)
-    end
-  end
-
-  describe "with block conditions" do
-    before(:each) do
-      @mailing = MailyHerald::OneTimeMailing.new
-      @mailing.title = "Foobar"
-      @mailing.subject = "Foo"
-      @mailing.template = "Sample template"
-      @mailing.list = @list
-      @mailing.start_at = "user.created_at"
-      @mailing.conditions = Proc.new {|user| user.weekly_notifications}
-      @mailing.enable
-      @mailing.save!
-    end
-
-    after(:each) do
-      @mailing.destroy
-      @entity.reload
-      @entity.update_attribute(:weekly_notifications, true)
-    end
-
-    it "should deliver when positive" do
-      expect(@mailing.has_conditions_proc?).to be_truthy
-
-      @list.subscribe!(@entity)
-      @mailing.set_schedules
-
-      expect(@mailing.schedules.for_entity(@entity).count).to eq(1)
-
-      schedule = @mailing.schedules.for_entity(@entity).last
-
-      expect(schedule.processing_at.to_i).to eq(@entity.created_at.to_i)
-      expect(@entity.weekly_notifications).to be_truthy
-      expect(@mailing.conditions_met?(@entity)).to be_truthy
-
-      @mailing.run
-
-      schedule.reload
-      expect(schedule.status).to eq(:delivered)
-    end
-
-    it "should skip when negative" do
-      expect(@mailing.has_conditions_proc?).to be_truthy
-
-      @list.subscribe!(@entity)
-      @mailing.set_schedules
-
-      expect(@mailing.schedules.for_entity(@entity).count).to eq(1)
-
-      schedule = @mailing.schedules.for_entity(@entity).last
-
-      expect(schedule.processing_at.to_i).to eq(@entity.created_at.to_i)
-
-      @entity.update_attribute(:weekly_notifications, false)
-
-      expect(@entity.weekly_notifications).to be_falsey
-      expect(@mailing.conditions_met?(@entity)).to be_falsey
-
-      @mailing.run
-
-      schedule.reload
-      expect(schedule.status).to eq(:skipped)
-    end
-  end
 end
